third_party/WebKit/LayoutTests/bluetooth/server/getPrimaryService/gen-no-permission-present-service.html

<!-- Generated by //third_party/WebKit/LayoutTests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../../resources/bluetooth/web-bluetooth-test.js"></script>
<script src="../../../resources/mojo-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  let expected = new DOMException('Origin is not allowed to access the ' +
                                  'service. Tip: Add the service UUID to ' +
                                  '\'optionalServices\' in requestDevice() ' +
                                  'options. https://goo.gl/HxfxSQ',
                                  'SecurityError');
  return setBluetoothFakeAdapter('HeartRateAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['heart_rate']}]}))
    .then(device => device.gatt.connect())
    .then(gatt => Promise.all([
      assert_promise_rejects_with_message(
        gatt.getPrimaryService(generic_access.alias), expected),
      assert_promise_rejects_with_message(
        gatt.getPrimaryService(generic_access.name), expected),
      assert_promise_rejects_with_message(
        gatt.getPrimaryService(generic_access.uuid), expected)]));
}, 'Request for present service without permission. Reject with SecurityError.');

</script>
